% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Genodata1}
\alias{Genodata1}
\title{Name of Data}
\format{
A data frame with xxxxx rows and xxxxx variables:
\describe{

...
}
}
\source{
\url{your_URL_}
}
\usage{
Genodata1
}
\description{
A dataset containing .......
}
\keyword{datasets}
