% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace_Ped.R
\name{Laplace_Ped}
\alias{Laplace_Ped}
\title{Laplace_Ped}
\usage{
Laplace_Ped(Data,Prior=c(1/2,1/2),Pop.col,Geno.cols,Pedigree)
}
\arguments{
\item{Data}{A data frame or matrix containing genotypic data from m markers as well as a column
indicating the subpopulation to which the individual belongs.Subpopulations must be coded
using consecutive integer numbers starting from 1.}

\item{Prior}{A vector of dimension 2 or a matrix of dimension (number of subpopulations) X 2,
containing the values of model hyperpameters (positive real numbers). The default value is (1/2,1/2).}

\item{Pop.col}{Integer indicating the column that contains the subpopulation each individual belongs to.}

\item{Geno.cols}{Vector containing the columns corresponding to genotypes in the input dataset.}

\item{Pedigree}{Data frame or matrix with three columns correspondint to: individual, sire and dam.
It must be ordered as the Data set.}
}
\value{
A list containing the Following elements:
Posterior_Means: a vector of dimension number of markers by 1 with entries
corresponding to the Laplace approximation to posterior means of the Fst for each marker
under the 'full model'.
N_Founders: A table with the number of founders in each subpopulation.
}
\description{
Estimates the posterior means of the Fst parameter using the Laplace
approximation under the full model proposed by Gianola et al. (2010) to infer selection
signatures using genomic data from diploid individuals and the likelihood function
derived by Martínez et al. (2017) to consider pedigree information.
}
\examples{
Data=Data2
Ex=Laplace_Ped(Data,Prior=c(1/2,1/2),Pop.col=4,Geno.cols=c(5:1004),Pedigree=Data2[,1:3])
summary(Ex$Posterior_Means)
}
\references{
Gianola, D., Simianer, H., Qanbari, S. 2010. A two-step method for detecting selection
signatures using genetic markers. Genetic Research Cambridge, 92; 141-155.

Martínez, C.A., Khare, K., Banerjee, A., Elzo, M.A. 2017.
Joint genome-wide prediction in several populations accounting for randomness
of genotypes: A hierarchical Bayes approach. I: Multivariate Gaussian priors
for marker effects and derivation of the joint probability mass function of genotypes.
Journal of Theoretical Biology 417, 8-19.
}
\author{
Carlos Alberto Martínez Niño (cmartinez@agrosavia.co).
}
