% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostSamp_Neutral.R
\name{PostSamp_Neutral}
\alias{PostSamp_Neutral}
\title{PostSamp_Neutral}
\usage{
PostSamp_Neutral(Data,Prior=c(1/2,1/2),N.Samples,Pop.col,Geno.cols)
}
\arguments{
\item{Data}{A data frame or matrix containing genotypic data from m markers as well as a column
indicating the subpopulation to which the individual belongs.
Subpopulations must be coded using consecutive integer numbers starting from 1.}

\item{Prior}{A vector of dimension 2 containing the values of model hyperpameters
(positive real numbers). The default value is (1/2,1/2).}

\item{N.Samples}{Integer corresponding to the number of samples used to perform the Monte Carlo
estimation of the Fst parameter.}

\item{Pop.col}{Integer indicating the column that contains the subpopulation each individual belongs to.}

\item{Geno.cols}{Vector containing the columns corresponding to genotypes in the input dataset.}
}
\value{
A list containing two elements. Theta.samples. Matrix of dimension (number of markers) X N.Samples
containing each sample used to compute the Monte Carlo estimate of Fst. PosteriorMeans. Vector of
dimension equal to the number of markers containing the Monte Carlo estimates of Fst under the
'null model'.
}
\description{
Estimates the Fst parameter using the 'null model' under the approach developed
by Gianola et al. (2010) to infer selection signatures using genomic data from diploid individuals
}
\examples{
Data=Data1
Ex1=PostSamp_Neutral(Data,Prior=c(1/2,1/3),N.Samples=5000,Pop.col=1,Geno.cols= c(2:ncol(Data)))
summary(Ex1$Posterior_Means)
}
\references{
Gianola, D., Simianer, H., Qanbari, S. 2010. A two-step method for detecting selection
signatures using genetic markers. Genetic Research Cambridge, 92; 141-155.
}
\author{
Carlos Alberto Martínez Niño (cmartinez@agrosavia.co).
}
