% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostSampPed.R
\name{PostSampPed}
\alias{PostSampPed}
\title{PostSampPed}
\usage{
PostSampPed(Data,Prior=c(1/2,1/2),N.Samples,Pop.col,Geno.cols,Pedigree)
}
\arguments{
\item{Data}{A data frame or matrix containing genotypic data from m markers as well as a column
indicating the subpopulation to which the individual belongs.
Subpopulations must be coded using consecutive integer numbers starting from 1.}

\item{Prior}{A vector of dimension 2 or a matrix of dimension (number of subpopulations) X 2,
containing the values of model hyperpameters (positive real numbers). The default value is (1/2,1/2).}

\item{N.Samples}{Integer corresponding to the number of samples used to perform the Monte Carlo
estimation of the Fst parameter.}

\item{Pop.col}{Integer indicating the column that contains the subpopulation each individual belongs to.}

\item{Geno.cols}{Vector containing the columns corresponding to genotypes in the input dataset.}

\item{Pedigree}{A data frame with four columns corresponding to Population, individual, sire and dam.
with missing values for unknown parents. All four variables have to be coded using positive integers.}
}
\value{
A list containing two elements:
Theta.samples: A matrix of dimension (number of markers) X N.Samples
containing each sample used to compute the Monte Carlo estimate of Fst.
PosteriorMeans: Vector of dimension equal to the number of markers containing
the Monte Carlo estimates of Fst under the'full model'.
}
\description{
Estimates the Fst parameter using an extension of the 'full model' under the approach developed
by Gianola et al. (2010) to infer selection signatures using genomic data from diploid individuals
incorporating pedigree information by using a modification of the likelihood derived by Mart?nez at al. (2017)
}
\examples{
Data=sim.1data
Ex1=PostSampPed(Data,Prior=c(1/2,1/3),N.Samples=5000,Pop.col=1,Geno.cols=c(5:ncol(Genodata1)),Pedigree=Data[,1:4])
summary(Ex1$Posterior_Means)
}
\references{
Gianola, D., Simianer, H., Qanbari, S. 2010. A two-step method for detecting selection
signatures using genetic markers. Genetic Research Cambridge, 92; 141-155.

Mart?nez, C.A., Khare, K., Banerjee, A., Elzo, M.A. 2017.
Joint genome-wide prediction in several populations accounting for randomness
of genotypes: A hierarchical Bayes approach. I: Multivariate Gaussian priors
for marker effects and derivation of the joint probability mass function of genotypes.
Journal of Theoretical Biology 417, 8-19.
}
\author{
Carlos Alberto Martínez Niño (cmartinez@agrosavia.co).
}
