% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2.R
\name{run2}
\alias{run2}
\title{run2}
\usage{
run2(k,Postmeans)
}
\arguments{
\item{k}{Integer indicating the number of groups.}

\item{Postmeans}{Vector of dimension equal to the number of selected markers containing estimated
posterior means of Fst under the 'full model'.}
}
\value{
A list containing the summary of the finite mixture model for each number of groups.
}
\description{
Performs clustering of estimated Fst parameters using a finite mixture model as
described in Gianola et al. (2010).
}
\references{
Gianola, D., Simianer, H., Qanbari, S. 2010. A two-step method for detecting selection
signatures using genetic markers. Genetic Research Cambridge, 92; 141-155.
}
\author{
Carlos Alberto Martínez Niño (cmartinez@agrosavia.co).
}
