% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data2}
\alias{Data2}
\title{Genotypic data with Pedigree}
\format{
A data frame with 1250 rows and 1004 variables
}
\usage{
data(Data2)
}
\description{
A data frame or matrix containing genotypic
data from 1250 individuals as well as colums
indicating sire, dam and subpopulation
}
\author{
Carlos Alberto Martínez Niño (cmartinez@agrosavia.co).
}
\keyword{datasets}
