% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Genodata1}
\alias{Genodata1}
\title{Name of Data}
\format{
A data frame with 1247 rows and 50 variables:
\describe{

...
}
}
\source{
\url{your_URL_}
}
\usage{
Genodata1
}
\description{
A data frame or matrix containing genotypic
data from m markers as well as a column indicating the subpopulation
to which the individual belongs.
Subpopulations must be coded using
consecutive integer numbers starting from 1.
}
\keyword{datasets}
